refactor(tool-manager): centralize backend path abstractions
